refactor: Re-organize as proper src layout.  Use uv for dependency management.
