+++ pyproject.toml
@@ -1,0 +1,37 @@
+[project]
+name = "alarmpi"
+version = "0.1.0"
+description = "Alarm clock and radio combination for Raspberry Pi"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "adafruit-circuitpython-dht",
+    "feedparser",
+    "google-cloud-texttospeech",
+    "mutagen",
+    "num2words",
+    "pydub",
+    "PyQt5-sip",
+    "PyQt5==5.15.10",
+    "pyqtspinner",
+    "python-dateutil",
+    "pyyaml",
+    "requests"
+]
+
+[dependency-groups]
+dev = [
+    "flake8",
+    "freezegun",
+    "ipython",
+    "pytest",
+    "pytest-xvfb"
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/alarmpi"]
+
