--- requirements.in
@@ -1,18 +1,0 @@
-adafruit-circuitpython-dht
-feedparser
-flake8
-freezegun
-google-cloud-texttospeech
-ipython
-mutagen
-num2words
-pip-tools
-pydub
-PyQt5-sip
-PyQt5==5.15.10
-pyqtspinner
-pytest
-pytest-xvfb
-python-dateutil
-pyyaml
-requests
