--- requirements.txt
@@ -1,197 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.10
-# by the following command:
-#
-#    pip-compile
-#
-adafruit-blinka==8.27.0
-    # via
-    #   adafruit-circuitpython-busdevice
-    #   adafruit-circuitpython-dht
-    #   adafruit-circuitpython-requests
-    #   adafruit-circuitpython-typing
-adafruit-circuitpython-busdevice==5.2.6
-    # via adafruit-circuitpython-typing
-adafruit-circuitpython-dht==4.0.3
-    # via -r requirements.in
-adafruit-circuitpython-requests==2.0.2
-    # via adafruit-circuitpython-typing
-adafruit-circuitpython-typing==1.9.5
-    # via
-    #   adafruit-blinka
-    #   adafruit-circuitpython-busdevice
-adafruit-platformdetect==3.57.0
-    # via adafruit-blinka
-adafruit-pureio==1.1.11
-    # via adafruit-blinka
-asttokens==2.4.1
-    # via stack-data
-build==1.0.3
-    # via pip-tools
-cachetools==5.3.2
-    # via google-auth
-certifi==2023.11.17
-    # via requests
-charset-normalizer==3.3.2
-    # via requests
-click==8.1.7
-    # via pip-tools
-decorator==5.1.1
-    # via ipython
-docopt==0.6.2
-    # via num2words
-exceptiongroup==1.2.0
-    # via
-    #   ipython
-    #   pytest
-executing==2.0.1
-    # via stack-data
-feedparser==6.0.10
-    # via -r requirements.in
-flake8==6.1.0
-    # via -r requirements.in
-freezegun==1.3.1
-    # via -r requirements.in
-google-api-core[grpc]==2.15.0
-    # via google-cloud-texttospeech
-google-auth==2.25.1
-    # via google-api-core
-google-cloud-texttospeech==2.15.0
-    # via -r requirements.in
-googleapis-common-protos==1.62.0
-    # via
-    #   google-api-core
-    #   grpcio-status
-grpcio==1.60.0
-    # via
-    #   google-api-core
-    #   grpcio-status
-grpcio-status==1.60.0
-    # via google-api-core
-idna==3.6
-    # via requests
-iniconfig==2.0.0
-    # via pytest
-ipython==8.18.1
-    # via -r requirements.in
-jedi==0.19.1
-    # via ipython
-matplotlib-inline==0.1.6
-    # via ipython
-mccabe==0.7.0
-    # via flake8
-mutagen==1.47.0
-    # via -r requirements.in
-num2words==0.5.13
-    # via -r requirements.in
-packaging==23.2
-    # via
-    #   build
-    #   pytest
-parso==0.8.3
-    # via jedi
-pexpect==4.9.0
-    # via ipython
-pip-tools==7.3.0
-    # via -r requirements.in
-pluggy==1.3.0
-    # via pytest
-prompt-toolkit==3.0.41
-    # via ipython
-proto-plus==1.23.0
-    # via google-cloud-texttospeech
-protobuf==4.25.1
-    # via
-    #   google-api-core
-    #   google-cloud-texttospeech
-    #   googleapis-common-protos
-    #   grpcio-status
-    #   proto-plus
-ptyprocess==0.7.0
-    # via pexpect
-pure-eval==0.2.2
-    # via stack-data
-pyasn1==0.5.1
-    # via
-    #   pyasn1-modules
-    #   rsa
-pyasn1-modules==0.3.0
-    # via google-auth
-pycodestyle==2.11.1
-    # via flake8
-pydub==0.25.1
-    # via -r requirements.in
-pyflakes==3.1.0
-    # via flake8
-pyftdi==0.55.0
-    # via adafruit-blinka
-pygments==2.17.2
-    # via ipython
-pyproject-hooks==1.0.0
-    # via build
-pyqt5==5.15.10
-    # via
-    #   -r requirements.in
-    #   pyqtspinner
-pyqt5-qt5==5.15.2
-    # via pyqt5
-pyqt5-sip==12.13.0
-    # via
-    #   -r requirements.in
-    #   pyqt5
-pyqtspinner==2.0.0
-    # via -r requirements.in
-pyserial==3.5
-    # via pyftdi
-pytest==7.4.3
-    # via
-    #   -r requirements.in
-    #   pytest-xvfb
-pytest-xvfb==3.0.0
-    # via -r requirements.in
-python-dateutil==2.8.2
-    # via
-    #   -r requirements.in
-    #   freezegun
-pyusb==1.2.1
-    # via pyftdi
-pyvirtualdisplay==3.0
-    # via pytest-xvfb
-pyyaml==6.0.1
-    # via -r requirements.in
-requests==2.31.0
-    # via
-    #   -r requirements.in
-    #   google-api-core
-rsa==4.9
-    # via google-auth
-sgmllib3k==1.0.0
-    # via feedparser
-six==1.16.0
-    # via
-    #   asttokens
-    #   python-dateutil
-stack-data==0.6.3
-    # via ipython
-tomli==2.0.1
-    # via
-    #   build
-    #   pip-tools
-    #   pyproject-hooks
-    #   pytest
-traitlets==5.14.0
-    # via
-    #   ipython
-    #   matplotlib-inline
-typing-extensions==4.8.0
-    # via adafruit-circuitpython-typing
-urllib3==2.1.0
-    # via requests
-wcwidth==0.2.12
-    # via prompt-toolkit
-wheel==0.42.0
-    # via pip-tools
-
-# The following packages are considered to be unsafe in a requirements file:
-# pip
-# setuptools
