+++ src/alarmpi/core/applugin.py
