+++ src/alarmpi/handlers/get_festival_tts.py
@@ -4,7 +4,7 @@ import tempfile
 import pydub
 import pydub.playback
 
-from src import aptts
+from alarmpi.core import aptts
 
 
 class FestivalTTSManager(aptts.AlarmpiTTS):
