+++ src/alarmpi/handlers/get_weather.py
@@ -8,7 +8,7 @@ import logging
 import requests
 from requests.exceptions import RequestException
 
-from src import apcontent
+from alarmpi.core import apcontent
 
 
 event_logger = logging.getLogger("eventLogger")
